refactor(HotBoardSelectionStrategies): 패키지 이동
